migrating to vue.js (incomplete)
