+++ hydroearth/data/datastore.py
@@ -31,7 +31,7 @@ def from_datastore(entity):
     return entity
 
 
-def list(limit=10, cursor=None):
+def list(limit=100, cursor=None):
     ds = get_client()
 
     query = ds.query(kind='Model')
@@ -73,6 +73,11 @@ def list_by_user(user_id, limit=10, cursor=None):
 
 
 # [END list_by_user]
+
+def count(id):
+    ds = get_client()
+
+    return -1
 
 
 def read(id):
