+++ requirements.txt
@@ -8,4 +8,5 @@ gunicorn==19.7.1
 oauth2client==4.1.2
 six==1.11.0
 requests[security]==2.18.4
-psq==0.6.0
+psq==0.7.0
+flasgger==0.9.1
